Refactor Cache Internal Classes
In #89, we introduced the Subscription API and split the Client I/F and Cache implementation into two:

* SwrClient - SwrCache
* SwrClientPlus - SwrCachePlus

Previously, SwrCache and SwrCachePlus had a direct inheritance relationship. This refactor extracts the internal logic and removes the inheritance.

With this change, the soil-query-testing library can now define the same extension functions for both:

```
val testClient1 = SwrCache(..).test()
val testClient2 = SwrCachePlus(..).test()
```
