+++ soil-query-core/src/commonMain/kotlin/soil/query/annotation/InternalSoilQueryApi.kt
@@ -1,0 +1,15 @@
+// Copyright 2024 Soil Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package soil.query.annotation
+
+@RequiresOptIn(message = "This API is internal to Soil Query and should not be used from outside the library.")
+@Retention(AnnotationRetention.BINARY)
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.TYPEALIAS,
+    AnnotationTarget.CONSTRUCTOR
+)
+annotation class InternalSoilQueryApi
