+++ soil-query-core/src/commonMain/kotlin/soil/query/core/NetworkConnectivity.kt
@@ -4,8 +4,15 @@
 package soil.query.core
 
 import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.filter
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.scan
+import kotlin.time.Duration
 
 /**
  * Interface for receiving events of network connectivity.
@@ -70,4 +77,17 @@ enum class NetworkConnectivityEvent {
      * The network is lost.
      */
     Lost
+}
+
+internal suspend fun observeOnNetworkReconnect(
+    networkConnectivity: NetworkConnectivity,
+    networkResumeAfterDelay: Duration,
+    collector: FlowCollector<Unit>
+) {
+    networkConnectivity.asFlow()
+        .distinctUntilChanged()
+        .scan(NetworkConnectivityEvent.Available to NetworkConnectivityEvent.Available) { acc, state -> state to acc.first }
+        .filter { it.first == NetworkConnectivityEvent.Available && it.second == NetworkConnectivityEvent.Lost }
+        .onEach { delay(networkResumeAfterDelay) }
+        .collect { collector.emit(Unit) }
 }
